fix : fix constructor and type

+++ src/main/java/gift/repository/CategoryRepository.java
@@ -1,6 +1,5 @@
 package gift.repository;
 
-import gift.domain.CategoryDTO;
 import gift.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
@@ -8,7 +7,7 @@ import java.util.List;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
     List<Category> findAll();
-    Category findById(long id);
-    void deleteById(long id);
+    Category findById(int id);
+    void deleteById(int id);
     Category save(Category category);
 }
